pyproject.toml
[project]
name = "auto-doc-ai"
version = "0.1.0"
description = "Synthetic DMV title document generation and local OCR/LayoutLMv3 experimentation toolkit"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.7.0",
    "faker>=40.8.0",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pillow>=11.0.0",
    "requests>=2.32.5",
    "pytesseract>=0.3.13",
    "seqeval>=1.2.2",
    "setuptools>=82.0.1",
    "svgwrite>=1.4.3",
    "torch>=2.10.0",
    "transformers>=5.3.0",
]
